+++ package.json
@@ -8,7 +8,7 @@
   "dependencies": {
     "react": "16.5.2",
     "react-dom": "16.5.2",
-    "@twilio/flex-plugin-scripts": "5.1.2"
+    "@twilio/flex-plugin-scripts": "6.0.3"
   },
   "devDependencies": {
     "@twilio/flex-ui": "^1",
