--- src/actions.js
@@ -1,1 +1,0 @@
-//Move Flex Actions here
